+++ NEWS.md
@@ -2,6 +2,7 @@
 * Greatly improved speed of segmentation model by adding internal function summarize1 in C++ and updated get_summary_stats().
 * Updated shiny_tracks to improve responsiveness to changes in the time window.
 * Updated traceplot() to automatically determine the number of MCMC iterations and add a line denoting the burn-in period.
+* Updated progress bar in segment_behavior() to prevent it from disappearing.
 
 # bayesmove 0.2.0 (2021-04-26)